requirements.txt
algopy=0.5.7=pypi_0
altair=4.1.0=py_1
anyio=3.5.0=py39haa95532_0
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py39h2bbff1b_0
arrow-cpp=8.0.0=py39hbd6f097_0
asttokens=2.0.5=pyhd3eb1b0_0
attrs=21.4.0=pyhd3eb1b0_0
aws-c-common=0.4.57=ha925a31_1
aws-c-event-stream=0.1.6=hd77b12b_5
aws-checksums=0.1.9=ha925a31_0
aws-sdk-cpp=1.8.185=hd77b12b_0
babel=2.9.1=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.11.1=py39haa95532_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
blinker=1.4=py39haa95532_0
boost-cpp=1.73.0=h2bbff1b_12
bottleneck=1.3.5=py39h080aedc_0
branca=0.5.0=pypi_0
brotli=1.0.9=h2bbff1b_7
brotli-bin=1.0.9=h2bbff1b_7
brotlipy=0.7.0=py39h2bbff1b_1003
bzip2=1.0.8=he774522_0
c-ares=1.18.1=h2bbff1b_0
ca-certificates=2022.07.19=haa95532_0
cachetools=4.2.2=pyhd3eb1b0_0
certifi=2022.6.15=py39haa95532_0
cffi=1.15.1=py39h2bbff1b_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py39haa95532_0
colorama=0.4.5=py39haa95532_0
commonmark=0.9.1=pyhd3eb1b0_0
console_shortcut=0.1.1=4
cryptography=37.0.1=py39h21b164f_0
cycler=0.11.0=pyhd8ed1ab_0
dataclasses=0.8=pyh6d0b6a4_7
debugpy=1.5.1=py39hd77b12b_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
docopt=0.6.2=pypi_0
emcee=3.1.2=pyhd8ed1ab_0
entrypoints=0.4=py39haa95532_0
executing=0.8.3=pyhd3eb1b0_0
flask=2.2.2=pypi_0
folium=0.12.1.post1=pypi_0
fonttools=4.36.0=py39hb82d6ee_0
freetype=2.10.4=hd328e21_0
future=0.18.2=py39haa95532_1
gflags=2.2.2=ha925a31_0
gitdb=4.0.7=pyhd3eb1b0_0
gitpython=3.1.18=pyhd3eb1b0_1
glog=0.5.0=hd77b12b_0
icc_rt=2019.0.0=h0cc432a_1
idna=3.3=pyhd3eb1b0_0
importlib-metadata=4.11.3=py39haa95532_0
intel-openmp=2021.4.0=haa95532_3556
ipykernel=6.9.1=py39haa95532_0
ipython=8.4.0=py39haa95532_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.5=pyhd3eb1b0_1
itsdangerous=2.1.2=pypi_0
jedi=0.18.1=py39haa95532_1
jinja2=3.0.3=pyhd3eb1b0_0
jpeg=9e=h2bbff1b_0
json5=0.9.6=pyhd3eb1b0_0
jsonschema=4.4.0=py39haa95532_0
kiwisolver=1.4.4=py39h2e07f2f_0
lerc=3.0=hd77b12b_0
libboost=1.73.0=h6c2663c_12
libbrotlicommon=1.0.9=h2bbff1b_7
libbrotlidec=1.0.9=h2bbff1b_7
libbrotlienc=1.0.9=h2bbff1b_7
libcurl=7.84.0=h86230a5_0
libdeflate=1.8=h2bbff1b_5
libpng=1.6.37=h2a8f88b_0
libprotobuf=3.20.1=h23ce68f_0
libsodium=1.0.18=h62dcd97_0
libssh2=1.10.0=hcd4344a_0
libthrift=0.15.0=he1d8c1a_0
libtiff=4.4.0=h8a3f274_0
libwebp=1.2.2=h2bbff1b_0
lz4-c=1.9.3=h2bbff1b_1
markupsafe=2.1.1=py39h2bbff1b_0
matplotlib-base=3.5.3=py39h581301d_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mistune=0.8.4=py39h2bbff1b_1000
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py39h2bbff1b_0
mkl_fft=1.3.1=py39h277e83a_0
mkl_random=1.2.2=py39hf11a4ad_0
munkres=1.1.4=pyh9f0ad1d_0
nbclassic=0.3.5=pyhd3eb1b0_0
nbclient=0.5.13=py39haa95532_0
nbconvert=6.4.4=py39haa95532_0
nbformat=5.3.0=py39haa95532_0
nest-asyncio=1.5.5=py39haa95532_0
notebook=6.4.12=py39haa95532_0
numdifftools=0.9.40=pypi_0
numexpr=2.8.3=py39hb80d3ca_0
numpy=1.23.1=py39h7a0a035_0
numpy-base=1.23.1=py39hca35cd5_0
openssl=1.1.1q=h2bbff1b_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.3=py39hd77b12b_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
patsy=0.5.2=pypi_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.2.0=py39hdc2b20a_1
pip=22.1.2=py39haa95532_0
pipreqs=0.4.11=pypi_0
prometheus_client=0.14.1=py39haa95532_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
protobuf=3.20.1=py39hd77b12b_0
pure_eval=0.2.2=pyhd3eb1b0_0
pyarrow=8.0.0=py39h953b917_0
pycparser=2.21=pyhd3eb1b0_0
pydeck=0.7.1=py39haa95532_0
pyextremes=2.2.4=pyhd8ed1ab_0
pygments=2.11.2=pyhd3eb1b0_0
pympler=0.9=py_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.4=pyhd3eb1b0_0
pyproj=3.3.1=pypi_0
pyrsistent=0.18.0=py39h196d8e1_0
pysocks=1.7.1=py39haa95532_0
python=3.9.12=h6244533_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.15.1=pyhd3eb1b0_0
python_abi=3.9=2_cp39
pytz=2022.1=py39haa95532_0
pywin32=302=py39h2bbff1b_2
pywinpty=2.0.2=py39h5da7b33_0
pyyaml=6.0=py39h2bbff1b_1
pyzmq=23.2.0=py39hd77b12b_0
re2=2022.04.01=hd77b12b_0
requests=2.28.1=py39haa95532_0
rich=12.5.1=py39haa95532_0
scipy=1.7.3=py39h0a974cb_0
semver=2.13.0=pyhd3eb1b0_0
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=61.2.0=py39haa95532_0
six=1.16.0=pyhd3eb1b0_1
smmap=4.0.0=pyhd3eb1b0_0
snappy=1.1.9=h6c2663c_0
sniffio=1.2.0=py39haa95532_1
soupsieve=2.3.1=pyhd3eb1b0_0
sqlite=3.39.2=h2bbff1b_0
stack_data=0.2.0=pyhd3eb1b0_0
statsmodels=0.13.2=pypi_0
streamlit=1.11.0=py39haa95532_0
streamlit-folium=0.6.14=pypi_0
terminado=0.13.1=py39haa95532_0
testpath=0.6.0=py39haa95532_0
tk=8.6.12=h2bbff1b_0
toml=0.10.2=pyhd3eb1b0_0
toolz=0.11.2=pyhd3eb1b0_0
tornado=6.1=py39h2bbff1b_0
tqdm=4.64.0=pyhd8ed1ab_0
traitlets=5.1.1=pyhd3eb1b0_0
typing-extensions=4.3.0=py39haa95532_0
typing_extensions=4.3.0=py39haa95532_0
tzdata=2022a=hda174b7_0
tzlocal=2.1=py39haa95532_0
unicodedata2=14.0.0=py39hb82d6ee_1
urllib3=1.26.11=py39haa95532_0
utf8proc=2.6.1=h2bbff1b_0
validators=0.18.2=pyhd3eb1b0_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
watchdog=2.1.6=py39haa95532_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py39haa95532_1
websocket-client=0.58.0=py39haa95532_4
werkzeug=2.2.2=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=3.5.2=py39haa95532_0
win_inet_pton=1.1.0=py39haa95532_0
wincertstore=0.2=py39haa95532_2
windrose=1.7.0=pypi_0
winpty=0.4.3=4
xz=5.2.5=h8cc25b3_1
yaml=0.2.5=he774522_0
yarg=0.1.9=pypi_0
zeromq=4.3.4=hd77b12b_0
zipp=3.8.0=py39haa95532_0
zlib=1.2.12=h8cc25b3_2
zstd=1.5.2=h19a0ad4_0
